// src/app/api/models/ticket-summary.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { UserSummary } from '../models/user-summary';
export interface TicketSummary {
  assignedTo?: UserSummary;
  category?: string;
  createdAt?: string;
  id?: string;
  priority?: 'LOW' | 'MEDIUM' | 'HIGH' | 'CRITICAL';
  status?: 'OPEN' | 'IN_PROGRESS' | 'RESOLVED' | 'CLOSED';
  title?: string;
}